requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
requests
python-dotenv
loguru
re
pydantic-settings